requirements.txt
pywin32==306
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.23
pytz==2023.3